+++ VersionHistory.md
@@ -1,8 +1,12 @@
 ## Version History
 
+### Version 1.0.3.0
+#### Enhancements:
+1. [Record Type Themes](https://github.com/shriprem/FWDataViz/blob/master/docs/file_type_extract_dialog.md)
+
 ### Version 1.0.2.0
 #### Enhancements:
-1. Automatic Detection of File Types
+1. [Automatic Detection of File Types](https://github.com/shriprem/FWDataViz/blob/master/docs/auto_detect_file_type.md)
 
 ### Version 1.0.1.0
 #### Enhancements: